Work through end of Chapter 5
This chapter is focused on building the AST printer for a
set of expression classes. It does so by generating a
series of expression classes using a separate java program
and then implementing the visitor pattern for each expression
type.

I noted in comments in the code why I think Rust does this
just fine using `impl`/`enum`, and I think that the resulting
code is super simple to write and modify to not warrant the
metaprogramming or visitor overhead. Eventually, I think
this solution might be replaced by a macro that generates
different structs instead of enums and implements each trait
for them, maybe with `!#[derive(Printer)]`, but I don't know
enough about rust to be sure if that's the way to go yet.