Work through chapter 6.3.3
Getting through most of the parser grammar here, in other
words working through how we turn the list of tokens into
actual expression data structures.

Got the very first bit of error handling in there too, since
we want to be able to recover from certain kinds of failures.

The next stage is going to be working on the winding back of
the stack whenever we find an error.