Work through chapter 6
The language can now parse expressions, but not statements.

I added two more example files that are just expressions. When
I did that I found a bug in my number parsing code that meant
it could only parse numbers of greater than one character.

I also added the parser to the main `run` loop, which is how I
found the bugs. There will be lots of room to improve the error
handline gode here because right now it just branches, but in
the future it could handl different errors differently (I think
this is what the next chapter is about).

Lox also tracks column number now which makes debugging easier
and could eventually pass on to tokens, even including start
and end to make statements like Rust does where the error is
underlined.

Lox is passed into the Parser just like with the Scanner.
Just like the Scanner, it could be passed in on creation and
then we could shuttle it through to each phase.

Finally, I updated the `Printer` trait for Expression so that
it just prints the expression whenever it's grouped so that
the expressions are printed as a lisp-like AST.