Fix comment formatting and error handling
Comment formatting is less of a big deal for the language semantics or behavior,
but it does very much clean things up when reading the code.

The error handling is a little less clear-cut and so probably will still change
a bit, but also goes a long way to make the code a little more coherent, if not
more clean. I like the way this method of error handling ties the different
sections together, even if it isn't how things should be long-term, it's a step
in the right direction.