Rework function implementation
Hoo boy, okay this is a doozie. This is the first time that I've
had to drastically separate from the book implementation. The
thing is, we just can't do 'Object' as a type like you can with
Java, so instead I'm improvising. I guess we'll see the implications
in the next few commits as to whether this is possible or not.

My biggeset worry is the mutable environment stuffed into the
function call object itself, also will we have to clone every
global variable to do this? Bummer, right?